requirements.txt
aiohttp==3.7.4
akinator.py==4.0.1
async-timeout==3.0.1
asyncode==1.0.0
attrs==20.3.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
discord==1.0.1
discord-webhook==0.11.0
discord.py==1.6.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gspread==3.6.0
httplib2==0.19.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
multidict==4.7.6
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==8.1.1
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.20
typing-extensions==3.7.4.3
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.26.3
yarl==1.5.1
